fix hdcs_open bug, rename Network_2 to Network, and cleanup some output
