+++ src/common/HDCS_REQUEST_HANDLER.h
@@ -3,7 +3,7 @@
 #define HDCS_REQUEST_HANDLER_H
 
 //#include "Network/client.h"
-#include "../Network_2/hdcs_networking.h"
+#include "../Network/hdcs_networking.h"
 #include "common/HDCS_REQUEST_CTX.h"
 
 namespace client {
